Finally done with scan_access_points.c.
Took a lot of work (first time writing C) but I got SSIDs to be printed
to screen. Only "bug" is upon an access point that doesn't broadcast an
SSID, a bunch of \x00 characters are printed instead of handling it
correctly. But that's out of scope for this example program.

Time to work on porting this to python.