Untested BytesIO, seq, consistent test_sendmsg C.
Started to implement converting Python objects to bytes() for
socket.send().

I don't know why I didn't need to port sequence code, started
implementing it.

Setting the sequence to 1 in the C code in the docstring of
test_sendmsg() made the output consistent.

Forgot to close sockets, not sure if pytest does it but I'll do it
explicitley just in case.