IT FINALLY WORKS THANK YOU JESUS
After WEEKS of working on this and refactoring basically everything to
bytearrays, it WORKS FUCK YEA.

On my Raspberry Pi, first test in test_ctrl.py passes. Data to and from
the kernel matches the C test.

Of course this doesn't work in Travis CI so I've disabled the test for
now. Since Travis-CI has no wifi adapters in their VMs, the nl80211
driver isn't loaded in the Linux kernel, so I get the NLE_OBJ_NOTFOUND
error code instead.

Need to figure out how to test this.